+++ app/assets/javascripts/init.js
@@ -81,6 +81,16 @@
 
     $('.port-bindings').appendablePortBindings();
 
+    $('.environment-variables dl.entries').editParameter({
+      fieldSelector: 'dt.variable-name label',
+      inputKey: 'input[data-index=name_%]'
+    });
+
+    $('.environment-variables dl.entries').editParameter({
+      fieldSelector: 'dd.variable-value span',
+      editSelector: 'dd.variable-value a.edit-action',
+      inputKey: 'input[data-index=value_%]'
+    });
     $('.journal-output').journalLoader();
 
     $('.journal-toggle').journalToggle();